requirements.txt
fastapi
fastapi[standard] # "fastapi[standard]"
pydantic-settingsexit
python-dotenv
sqlmodel
psycopg2
asyncpg
alembic
passlib
pyjwt[crypto]
